Move output snapshotting and fingerprint registry to execution
